pyproject.toml
[tool.poetry]
name = "plusone"
version = "0.1.0"
description = ""
authors = ["karim mattar <karimhady111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
black = "^24.4.2"
pylint = "^3.2.3"
drf-yasg = "^1.21.7"
pillow = "^10.3.0"
django-guardian = "^2.4.0"
django-filter = "^24.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
coverage = "^7.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
